sharing base cell editor options for the same language.
